Enhance order and payment processing features with Stripe integration; update order status management and improve API endpoints for user authentication and order handling

+++ backend/src/main/java/com/greata/cozy/dto/OrderResponseItemDTO.java
@@ -39,7 +39,7 @@ public class OrderResponseItemDTO {
         this.price = price;
     }
 
-    public int getQuantity() {
+    public long getQuantity() {
         return quantity;
     }
